split up image tests for the new query code
