+++ models/world_coordinates.py
@@ -117,13 +117,15 @@ class WorldCoordinates(Base, AutoIDMixin, FileOnDiskMixin, HasBitFlagBadness):
         from models.provenance import Provenance
 
         with (SmartSession(session) as session):
-            subs = session.scalars(
-                sa.select(Image).where(
-                    Image.provenance.has(Provenance.upstreams.any(Provenance.id == self.provenance.id)),
-                    Image.upstream_images.any(Image.id == self.sources.image_id),
-                )
-            ).all()
-            output = subs
+            output = []
+            if self.provenance is not None:
+                subs = session.scalars(
+                    sa.select(Image).where(
+                        Image.provenance.has(Provenance.upstreams.any(Provenance.id == self.provenance.id)),
+                        Image.upstream_images.any(Image.id == self.sources.image_id),
+                    )
+                ).all()
+                output += subs
 
             if siblings:
                 sources = session.scalars(sa.select(SourceList).where(SourceList.id == self.sources_id)).all()
