+++ tests/models/test_image.py
@@ -429,97 +429,6 @@ def test_image_enum_values(sim_image1):
                     os.rmdir(folder)
 
 
-def test_image_upstreams_downstreams(sim_image1, sim_reference, provenance_extra, data_dir):
-    with SmartSession() as session:
-        sim_image1 = sim_image1.merge_all(session)
-        sim_reference = sim_reference.merge_all(session)
-
-        # make sure the new image matches the reference in all these attributes
-        sim_image1.filter = sim_reference.filter
-        sim_image1.target = sim_reference.target
-        sim_image1.section_id = sim_reference.section_id
-
-        new = Image.from_new_and_ref(sim_image1, sim_reference.image)
-        new.provenance = session.merge(provenance_extra)
-
-        # save and delete at the end
-        cleanup = ImageCleanup.save_image(new)
-
-        session.add(new)
-        session.commit()
-
-    # new make sure a new session can find all the upstreams/downstreams
-    with SmartSession() as session:
-        sim_image1 = sim_image1.merge_all(session)
-        sim_reference = sim_reference.merge_all(session)
-        new = new.merge_all(session)
-
-        # check the upstreams/downstreams for the new image
-        upstream_ids = [u.id for u in new.get_upstreams(session=session)]
-        assert sim_image1.id in upstream_ids
-        assert sim_reference.image_id in upstream_ids
-        downstream_ids = [d.id for d in new.get_downstreams(session=session)]
-        assert len(downstream_ids) == 0
-
-        upstream_ids = [u.id for u in sim_image1.get_upstreams(session=session)]
-        assert [sim_image1.exposure_id] == upstream_ids
-        downstream_ids = [d.id for d in sim_image1.get_downstreams(session=session)]
-        assert [new.id] == downstream_ids  # should be the only downstream
-
-        # check the upstreams/downstreams for the reference image
-        upstreams = sim_reference.image.get_upstreams(session=session)
-        assert len(upstreams) == 5  # was made of five images
-        assert all([isinstance(u, Image) for u in upstreams])
-        source_images_ids = [im.id for im in sim_reference.image.upstream_images]
-        upstream_ids = [u.id for u in upstreams]
-        assert set(upstream_ids) == set(source_images_ids)
-        downstream_ids = [d.id for d in sim_reference.image.get_downstreams(session=session)]
-        assert [new.id] == downstream_ids  # should be the only downstream
-
-        # test for the Image.downstream relationship
-        assert len(upstreams[0].downstream_images) == 1
-        assert upstreams[0].downstream_images == [sim_reference.image]
-        assert len(upstreams[1].downstream_images) == 1
-        assert upstreams[1].downstream_images == [sim_reference.image]
-
-        assert len(sim_image1.downstream_images) == 1
-        assert sim_image1.downstream_images == [new]
-
-        assert len(sim_reference.image.downstream_images) == 1
-        assert sim_reference.image.downstream_images == [new]
-
-        assert len(new.downstream_images) == 0
-
-        # add a second "new" image using one of the reference's upstreams instead of the reference
-        new2 = Image.from_new_and_ref(sim_image1, upstreams[0])
-        new2.provenance = session.merge(provenance_extra)
-        new2.mjd += 1  # make sure this image has a later MJD, so it comes out later on the downstream list!
-
-        # save and delete at the end
-        cleanup2 = ImageCleanup.save_image(new2)
-
-        session.add(new2)
-        session.commit()
-
-        session.refresh(upstreams[0])
-        assert len(upstreams[0].downstream_images) == 2
-        assert set(upstreams[0].downstream_images) == set([sim_reference.image, new2])
-
-        session.refresh(upstreams[1])
-        assert len(upstreams[1].downstream_images) == 1
-        assert upstreams[1].downstream_images == [sim_reference.image]
-
-        session.refresh(sim_image1)
-        assert len(sim_image1.downstream_images) == 2
-        assert set(sim_image1.downstream_images) == set([new, new2])
-
-        session.refresh(sim_reference.image)
-        assert len(sim_reference.image.downstream_images) == 1
-        assert sim_reference.image.downstream_images == [new]
-
-        assert len(new2.downstream_images) == 0
-
-
 def test_image_preproc_bitflag( sim_image1 ):
 
     with SmartSession() as session:
@@ -554,458 +463,6 @@ def test_image_preproc_bitflag( sim_image1 ):
                                                                        image_preprocessing_inverse) )
                        == string_to_bitflag( 'overscan, fringe', image_preprocessing_inverse ) ) )
         assert q.count() == 0
-
-
-def test_image_badness(sim_image1):
-
-    with SmartSession() as session:
-        sim_image1 = session.merge(sim_image1)
-
-        # this is not a legit "badness" keyword...
-        with pytest.raises(ValueError, match='Keyword "foo" not recognized'):
-            sim_image1.badness = 'foo'
-
-        # this is a legit keyword, but for cutouts, not for images
-        with pytest.raises(ValueError, match='Keyword "cosmic ray" not recognized'):
-            sim_image1.badness = 'cosmic ray'
-
-        # this is a legit keyword, but for images, using no space and no capitalization
-        sim_image1.badness = 'brightsky'
-
-        # retrieving this keyword, we do get it capitalized and with a space:
-        assert sim_image1.badness == 'bright sky'
-        assert sim_image1.bitflag == 2 ** 5  # the bright sky bit is number 5
-
-        # what happens when we add a second keyword?
-        sim_image1.badness = 'Bright_sky, Banding'  # try this with capitalization and underscores
-        assert sim_image1.bitflag == 2 ** 5 + 2 ** 1  # the bright sky bit is number 5, banding is number 1
-        assert sim_image1.badness == 'banding, bright sky'
-
-        # now add a third keyword, but on the Exposure
-        sim_image1.exposure.badness = 'saturation'
-        session.add(sim_image1)
-        session.commit()
-
-        # a manual way to propagate bitflags downstream
-        sim_image1.exposure.update_downstream_badness(session=session)  # make sure the downstreams get the new badness
-        session.commit()
-        assert sim_image1.bitflag == 2 ** 5 + 2 ** 3 + 2 ** 1  # saturation bit is 3
-        assert sim_image1.badness == 'banding, saturation, bright sky'
-
-        # adding the same keyword on the exposure and the image makes no difference
-        sim_image1.exposure.badness = 'Banding'
-        sim_image1.exposure.update_downstream_badness(session=session)  # make sure the downstreams get the new badness
-        session.commit()
-        assert sim_image1.bitflag == 2 ** 5 + 2 ** 1
-        assert sim_image1.badness == 'banding, bright sky'
-
-        # try appending keywords to the image
-        sim_image1.append_badness('shaking')
-        assert sim_image1.bitflag == 2 ** 5 + 2 ** 2 + 2 ** 1  # shaking bit is 2
-        assert sim_image1.badness == 'banding, shaking, bright sky'
-
-
-def test_multiple_images_badness(
-        sim_image1,
-        sim_image2,
-        sim_image3,
-        sim_image5,
-        sim_image6,
-        provenance_extra
-):
-    try:
-        with SmartSession() as session:
-            sim_image1 = session.merge(sim_image1)
-            sim_image2 = session.merge(sim_image2)
-            sim_image3 = session.merge(sim_image3)
-            sim_image5 = session.merge(sim_image5)
-            sim_image6 = session.merge(sim_image6)
-
-            images = [sim_image1, sim_image2, sim_image3, sim_image5, sim_image6]
-            cleanups = []
-            filter = 'g'
-            target = str(uuid.uuid4())
-            project = 'test project'
-            for im in images:
-                im.filter = filter
-                im.target = target
-                im.project = project
-                session.add(im)
-
-            session.commit()
-
-            # the image itself is marked bad because of bright sky
-            sim_image2.badness = 'BrightSky'
-            assert sim_image2.badness == 'bright sky'
-            assert sim_image2.bitflag == 2 ** 5
-            session.commit()
-
-            # note that this image is not directly bad, but the exposure has banding
-            sim_image3.exposure.badness = 'banding'
-            sim_image3.exposure.update_downstream_badness(session=session)
-            session.commit()
-
-            assert sim_image3.badness == 'banding'
-            assert sim_image1._bitflag == 0  # the exposure is bad!
-            assert sim_image3.bitflag == 2 ** 1
-            session.commit()
-
-            # find the images that are good vs bad
-            good_images = session.scalars(sa.select(Image).where(Image.bitflag == 0)).all()
-            assert sim_image1.id in [i.id for i in good_images]
-
-            bad_images = session.scalars(sa.select(Image).where(Image.bitflag != 0)).all()
-            assert sim_image2.id in [i.id for i in bad_images]
-            assert sim_image3.id in [i.id for i in bad_images]
-
-            # make an image from the two bad exposures using subtraction
-
-            sim_image4 = Image.from_new_and_ref(sim_image3, sim_image2)
-            sim_image4.provenance = provenance_extra
-            sim_image4.provenance.upstreams = sim_image4.get_upstream_provenances()
-            cleanups.append(ImageCleanup.save_image(sim_image4))
-            sim_image4 = session.merge(sim_image4)
-            images.append(sim_image4)
-            session.commit()
-
-            assert sim_image4.id is not None
-            assert sim_image4.ref_image == sim_image2
-            assert sim_image4.new_image == sim_image3
-
-            # check that badness is loaded correctly from both parents
-            assert sim_image4.badness == 'banding, bright sky'
-            assert sim_image4._bitflag == 0  # the image itself is not flagged
-            assert sim_image4.bitflag == 2 ** 1 + 2 ** 5
-
-            # check that filtering on this value gives the right bitflag
-            bad_images = session.scalars(sa.select(Image).where(Image.bitflag == 2 ** 1 + 2 ** 5)).all()
-            assert sim_image4.id in [i.id for i in bad_images]
-            assert sim_image3.id not in [i.id for i in bad_images]
-            assert sim_image2.id not in [i.id for i in bad_images]
-
-            # check that adding a badness on the image itself is added to the total badness
-            sim_image4.badness = 'saturation'
-            session.add(sim_image4)
-            session.commit()
-            assert sim_image4.badness == 'banding, saturation, bright sky'
-            assert sim_image4._bitflag == 2 ** 3  # only this bit is from the image itself
-
-            # make a new subtraction:
-            sim_image7 = Image.from_ref_and_new(sim_image6, sim_image5)
-            sim_image7.provenance = provenance_extra
-            cleanups.append(ImageCleanup.save_image(sim_image7))
-            sim_image7 = session.merge(sim_image7)
-            images.append(sim_image7)
-            session.commit()
-
-            # check that the new subtraction is not flagged
-            assert sim_image7.badness == ''
-            assert sim_image7._bitflag == 0
-            assert sim_image7.bitflag == 0
-
-            good_images = session.scalars(sa.select(Image).where(Image.bitflag == 0)).all()
-            assert sim_image5.id in [i.id for i in good_images]
-            assert sim_image5.id in [i.id for i in good_images]
-            assert sim_image7.id in [i.id for i in good_images]
-
-            bad_images = session.scalars(sa.select(Image).where(Image.bitflag != 0)).all()
-            assert sim_image5.id not in [i.id for i in bad_images]
-            assert sim_image6.id not in [i.id for i in bad_images]
-            assert sim_image7.id not in [i.id for i in bad_images]
-
-            # let's try to coadd an image based on some good and bad images
-            # as a reminder, sim_image2 has bright sky (5),
-            # sim_image3's exposure has banding (1), while
-            # sim_image4 has saturation (3).
-
-            # make a coadded image (without including the subtraction sim_image4):
-            sim_image8 = Image.from_images([sim_image1, sim_image2, sim_image3, sim_image5, sim_image6])
-            sim_image8.provenance = provenance_extra
-            cleanups.append(ImageCleanup.save_image(sim_image8))
-            images.append(sim_image8)
-            sim_image8 = session.merge(sim_image8)
-            session.commit()
-
-            assert sim_image8.badness == 'banding, bright sky'
-            assert sim_image8.bitflag == 2 ** 1 + 2 ** 5
-
-            # does this work in queries (i.e., using the bitflag hybrid expression)?
-            bad_images = session.scalars(sa.select(Image).where(Image.bitflag != 0)).all()
-            assert sim_image8.id in [i.id for i in bad_images]
-            bad_coadd = session.scalars(sa.select(Image).where(Image.bitflag == 2 ** 1 + 2 ** 5)).all()
-            assert sim_image8.id in [i.id for i in bad_coadd]
-
-            # get rid of this coadd to make a new one
-            sim_image8.delete_from_disk_and_database(session=session)
-            cleanups.pop()
-            images.pop()
-
-            # now let's add the subtraction image to the coadd:
-            # make a coadded image (now including the subtraction sim_image4):
-            sim_image8 = Image.from_images([sim_image1, sim_image2, sim_image3, sim_image4, sim_image5, sim_image6])
-            sim_image8.provenance = provenance_extra
-            cleanups.append(ImageCleanup.save_image(sim_image8))
-            sim_image8 = session.merge(sim_image8)
-            images.append(sim_image8)
-            session.commit()
-
-            assert sim_image8.badness == 'banding, saturation, bright sky'
-            assert sim_image8.bitflag == 2 ** 1 + 2 ** 3 + 2 ** 5  # this should be 42
-
-            # does this work in queries (i.e., using the bitflag hybrid expression)?
-            bad_images = session.scalars(sa.select(Image).where(Image.bitflag != 0)).all()
-            assert sim_image8.id in [i.id for i in bad_images]
-            bad_coadd = session.scalars(sa.select(Image).where(Image.bitflag == 42)).all()
-            assert sim_image8.id in [i.id for i in bad_coadd]
-
-            # try to add some badness to one of the underlying exposures
-            sim_image1.exposure.badness = 'shaking'
-            session.add(sim_image1)
-            sim_image1.exposure.update_downstream_badness(session=session)
-            session.commit()
-
-            assert 'shaking' in sim_image1.badness
-            assert 'shaking' in sim_image8.badness
-
-    finally:  # cleanup
-        with SmartSession() as session:
-            session.autoflush = False
-            for im in images:
-                im = im.merge_all(session)
-                exp = im.exposure
-                im.delete_from_disk_and_database(session=session, commit=False)
-
-                if exp is not None and sa.inspect(exp).persistent:
-                    session.delete(exp)
-
-            session.commit()
-
-
-def test_image_coordinates():
-    image = Image('coordinates.fits', ra=None, dec=None, nofile=True)
-    assert image.ecllat is None
-    assert image.ecllon is None
-    assert image.gallat is None
-    assert image.gallon is None
-
-    image = Image('coordinates.fits', ra=123.4, dec=None, nofile=True)
-    assert image.ecllat is None
-    assert image.ecllon is None
-    assert image.gallat is None
-    assert image.gallon is None
-
-    image = Image('coordinates.fits', ra=123.4, dec=56.78, nofile=True)
-    assert abs(image.ecllat - 35.846) < 0.01
-    assert abs(image.ecllon - 111.838) < 0.01
-    assert abs(image.gallat - 33.542) < 0.01
-    assert abs(image.gallon - 160.922) < 0.01
-
-
-def test_image_cone_search( provenance_base ):
-    with SmartSession() as session:
-        image1 = None
-        image2 = None
-        image3 = None
-        image4 = None
-        try:
-            kwargs = { 'format': 'fits',
-                       'exp_time': 60.48,
-                       'section_id': 'x',
-                       'project': 'x',
-                       'target': 'x',
-                       'instrument': 'DemoInstrument',
-                       'telescope': 'x',
-                       'filter': 'r',
-                       'ra_corner_00': 0,
-                       'ra_corner_01': 0,
-                       'ra_corner_10': 0,
-                       'ra_corner_11': 0,
-                       'dec_corner_00': 0,
-                       'dec_corner_01': 0,
-                       'dec_corner_10': 0,
-                       'dec_corner_11': 0,
-                      }
-            image1 = Image(ra=120., dec=10., provenance=provenance_base, **kwargs )
-            image1.mjd = np.random.uniform(0, 1) + 60000
-            image1.end_mjd = image1.mjd + 0.007
-            clean1 = ImageCleanup.save_image( image1 )
-
-            image2 = Image(ra=120.0002, dec=9.9998, provenance=provenance_base, **kwargs )
-            image2.mjd = np.random.uniform(0, 1) + 60000
-            image2.end_mjd = image2.mjd + 0.007
-            clean2 = ImageCleanup.save_image( image2 )
-
-            image3 = Image(ra=120.0005, dec=10., provenance=provenance_base, **kwargs )
-            image3.mjd = np.random.uniform(0, 1) + 60000
-            image3.end_mjd = image3.mjd + 0.007
-            clean3 = ImageCleanup.save_image( image3 )
-
-            image4 = Image(ra=60., dec=0., provenance=provenance_base, **kwargs )
-            image4.mjd = np.random.uniform(0, 1) + 60000
-            image4.end_mjd = image4.mjd + 0.007
-            clean4 = ImageCleanup.save_image( image4 )
-
-            session.add( image1 )
-            session.add( image2 )
-            session.add( image3 )
-            session.add( image4 )
-
-            sought = session.query( Image ).filter( Image.cone_search(120., 10., rad=1.02) ).all()
-            soughtids = set( [ s.id for s in sought ] )
-            assert { image1.id, image2.id }.issubset( soughtids )
-            assert len( { image3.id, image4.id } & soughtids ) == 0
-
-            sought = session.query( Image ).filter( Image.cone_search(120., 10., rad=2.) ).all()
-            soughtids = set( [ s.id for s in sought ] )
-            assert { image1.id, image2.id, image3.id }.issubset( soughtids )
-            assert len( { image4.id } & soughtids ) == 0
-
-            sought = session.query( Image ).filter( Image.cone_search(120., 10., 0.017, radunit='arcmin') ).all()
-            soughtids = set( [ s.id for s in sought ] )
-            assert { image1.id, image2.id }.issubset( soughtids )
-            assert len( { image3.id, image4.id } & soughtids ) == 0
-
-            sought = session.query( Image ).filter( Image.cone_search(120., 10., 0.0002833, radunit='degrees') ).all()
-            soughtids = set( [ s.id for s in sought ] )
-            assert { image1.id, image2.id }.issubset( soughtids )
-            assert len( { image3.id, image4.id } & soughtids ) == 0
-
-            sought = session.query( Image ).filter( Image.cone_search(120., 10., 4.9451e-6, radunit='radians') ).all()
-            soughtids = set( [ s.id for s in sought ] )
-            assert { image1.id, image2.id }.issubset( soughtids )
-            assert len( { image3.id, image4.id } & soughtids ) == 0
-
-            sought = session.query( Image ).filter( Image.cone_search(60, -10, 1.) ).all()
-            soughtids = set( [ s.id for s in sought ] )
-            assert len( { image1.id, image2.id, image3.id, image4.id } & soughtids ) == 0
-
-            with pytest.raises( ValueError, match='.*unknown radius unit' ):
-                sought = Image.cone_search( 0., 0., 1., 'undefined_unit' )
-        finally:
-            for i in [ image1, image2, image3, image4 ]:
-                if ( i is not None ) and sa.inspect( i ).persistent:
-                    session.delete( i )
-            session.commit()
-
-
-# Really, we should also do some speed tests, but that
-# is outside the scope of the always-run tests.
-def test_four_corners( provenance_base ):
-
-    with SmartSession() as session:
-        image1 = None
-        image2 = None
-        image3 = None
-        image4 = None
-        try:
-            kwargs = { 'format': 'fits',
-                       'exp_time': 60.48,
-                       'section_id': 'x',
-                       'project': 'x',
-                       'target': 'x',
-                       'instrument': 'DemoInstrument',
-                       'telescope': 'x',
-                       'filter': 'r',
-                      }
-            # RA numbers are made ugly from cos(dec).
-            # image1: centered on 120, 40, square to the sky
-            image1 = Image( ra=120, dec=40.,
-                            ra_corner_00=119.86945927, ra_corner_01=119.86945927,
-                            ra_corner_10=120.13054073, ra_corner_11=120.13054073,
-                            dec_corner_00=39.9, dec_corner_01=40.1, dec_corner_10=39.9, dec_corner_11=40.1,
-                            provenance=provenance_base, nofile=True, **kwargs )
-            image1.mjd = np.random.uniform(0, 1) + 60000
-            image1.end_mjd = image1.mjd + 0.007
-            clean1 = ImageCleanup.save_image( image1 )
-
-            # image2: centered on 120, 40, at a 45° angle
-            image2 = Image( ra=120, dec=40.,
-                            ra_corner_00=119.81538753, ra_corner_01=120, ra_corner_11=120.18461247, ra_corner_10=120,
-                            dec_corner_00=40, dec_corner_01=40.14142136, dec_corner_11=40, dec_corner_10=39.85857864,
-                            provenance=provenance_base, nofile=True, **kwargs )
-            image2.mjd = np.random.uniform(0, 1) + 60000
-            image2.end_mjd = image2.mjd + 0.007
-            clean2 = ImageCleanup.save_image( image2 )
-
-            # image3: centered offset by (0.025, 0.025) linear arcsec from 120, 40, square on sky
-            image3 = Image( ra=120.03264714, dec=40.025,
-                            ra_corner_00=119.90210641, ra_corner_01=119.90210641,
-                            ra_corner_10=120.16318787, ra_corner_11=120.16318787,
-                            dec_corner_00=39.975, dec_corner_01=40.125, dec_corner_10=39.975, dec_corner_11=40.125,
-                            provenance=provenance_base, nofile=True, **kwargs )
-            image3.mjd = np.random.uniform(0, 1) + 60000
-            image3.end_mjd = image3.mjd + 0.007
-            clean3 = ImageCleanup.save_image( image3 )
-
-            # imagepoint and imagefar are used to test Image.containing and Image.find_containing,
-            # as Image is the only example of a SpatiallyIndexed thing we have so far.
-            # The corners don't matter for these given how they'll be used.
-            imagepoint = Image( ra=119.88, dec=39.95,
-                                ra_corner_00=-.001, ra_corner_01=0.001, ra_corner_10=-0.001,
-                                ra_corner_11=0.001, dec_corner_00=0, dec_corner_01=0, dec_corner_10=0, dec_corner_11=0,
-                                provenance=provenance_base, nofile=True, **kwargs )
-            imagepoint.mjd = np.random.uniform(0, 1) + 60000
-            imagepoint.end_mjd = imagepoint.mjd + 0.007
-            clearpoint = ImageCleanup.save_image( imagepoint )
-
-            imagefar = Image( ra=30, dec=-10,
-                              ra_corner_00=0, ra_corner_01=0, ra_corner_10=0,
-                              ra_corner_11=0, dec_corner_00=0, dec_corner_01=0, dec_corner_10=0, dec_corner_11=0,
-                              provenance=provenance_base, nofile=True, **kwargs )
-            imagefar.mjd = np.random.uniform(0, 1) + 60000
-            imagefar.end_mjd = imagefar.mjd + 0.007
-            clearfar = ImageCleanup.save_image( imagefar )
-
-            session.add( image1 )
-            session.add( image2 )
-            session.add( image3 )
-            session.add( imagepoint )
-            session.add( imagefar )
-
-            sought = session.query( Image ).filter( Image.containing( 120, 40 ) ).all()
-            soughtids = set( [ s.id for s in sought ] )
-            assert { image1.id, image2.id, image3.id }.issubset( soughtids )
-            assert len( { imagepoint.id, imagefar.id } & soughtids ) == 0
-
-            sought = session.query( Image ).filter( Image.containing( 119.88, 39.95 ) ).all()
-            soughtids = set( [ s.id for s in sought ] )
-            assert { image1.id }.issubset( soughtids  )
-            assert len( { image2.id, image3.id, imagepoint.id, imagefar.id } & soughtids ) == 0
-
-            sought = session.query( Image ).filter( Image.containing( 120, 40.12 ) ).all()
-            soughtids = set( [ s.id for s in sought ] )
-            assert { image2.id, image3.id }.issubset( soughtids )
-            assert len( { image1.id, imagepoint.id, imagefar.id } & soughtids ) == 0
-
-            sought = session.query( Image ).filter( Image.containing( 120, 39.88 ) ).all()
-            soughtids = set( [ s.id for s in sought ] )
-            assert { image2.id }.issubset( soughtids )
-            assert len( { image1.id, image3.id, imagepoint.id, imagefar.id } & soughtids ) == 0
-
-            sought = Image.find_containing( imagepoint, session=session )
-            soughtids = set( [ s.id for s in sought ] )
-            assert { image1.id }.issubset( soughtids )
-            assert len( { image2.id, image3.id, imagepoint.id, imagefar.id } & soughtids ) == 0
-
-            sought = session.query( Image ).filter( Image.containing( 0, 0 ) ).all()
-            soughtids = set( [ s.id for s in sought ] )
-            assert len( { image1.id, image2.id, image3.id, imagepoint.id, imagefar.id } & soughtids ) == 0
-
-            sought = Image.find_containing( imagefar, session=session )
-            soughtids = set( [ s.id for s in sought ] )
-            assert len( { image1.id, image2.id, image3.id, imagepoint.id, imagefar.id } & soughtids ) == 0
-
-            sought = session.query( Image ).filter( Image.within( image1 ) ).all()
-            soughtids = set( [ s.id for s in sought ] )
-            assert { image1.id, image2.id, image3.id, imagepoint.id }.issubset( soughtids )
-            assert len( { imagefar.id } & soughtids ) == 0
-
-            sought = session.query( Image ).filter( Image.within( imagefar ) ).all()
-            soughtids = set( [ s.id for s in sought ] )
-            assert len( { image1.id, image2.id, image3.id, imagepoint.id, imagefar.id } & soughtids ) == 0
-
-        finally:
-            session.rollback()
 
 
 def test_image_from_exposure(sim_exposure1, provenance_base):
